Use common view por input color picker
