+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/fill-color-by-value-view.js
@@ -3,7 +3,7 @@ var CoreView = require('backbone/core-view');
 var InputCollection = require('builder/components/input-collection/input-collection');
 var InputColorByValueView = require('builder/components/form-components/editors/fill-color/inputs/input-color-by-value');
 
-var FillTemplate = require('builder/components/form-components/editors/fill/fill-template.tpl');
+var FillTemplate = require('builder/components/input-fill/input-fill.tpl');
 var DialogModel = require('builder/components/dialog/dialog-model');
 var DialogView = require('builder/components/dialog/dialog-view');
 var PopupManager = require('builder/components/popup-manager');
