+++ lib/assets/javascripts/builder/components/form-components/editors/fill/fill.js
@@ -10,7 +10,7 @@ var InputCollection = require('./input-collection');
 var InputNumber = require('builder/components/input-number/input-number');
 var InputColor = require('builder/components/input-color/input-color');
 
-var template = require('./fill-template.tpl');
+var template = require('builder/components/input-fill/input-fill.tpl');
 
 var PopupManager = require('builder/components/popup-manager');
