+++ lib/assets/javascripts/builder/components/form-components/editors/stroke-color/stroke-color-view.js
@@ -1,12 +1,5 @@
-var CoreView = require('backbone/core-view');
-var DialogModel = require('builder/components/dialog/dialog-model');
-var DialogView = require('builder/components/dialog/dialog-view');
-
-var InputCollection = require('builder/components/input-collection/input-collection');
+var InputFillView = require('builder/components/input-fill/input-fill-view');
 var InputStrokeColor = require('builder/components/form-components/editors/stroke-color/inputs/input-stroke-color');
-
-var FillTemplate = require('builder/components/form-components/editors/fill/fill-template.tpl');
-var PopupManager = require('builder/components/popup-manager');
 
 var checkAndBuildOpts = require('builder/helpers/required-opts');
 
@@ -22,55 +15,17 @@ var REQUIRED_OPTS = [
   'strokeColorModel'
 ];
 
-module.exports = CoreView.extend({
-  className: 'Form-InputFill CDB-OptionInput CDB-Text js-input',
-
-  events: {
-    focus: function () {
-      this.trigger('focus', this);
-    },
-    blur: function () {
-      this.trigger('blur', this);
-    }
-  },
-
+module.exports = InputFillView.extend({
   initialize: function (options) {
     checkAndBuildOpts(options, REQUIRED_OPTS, this);
 
     this._initViews();
   },
 
-  render: function () {
-    this.clearSubViews();
-    this.$el.empty();
-    this._initViews();
-
-    return this;
-  },
-
-  _initViews: function () {
-    this.$el.append(FillTemplate());
-
-    if (this.options.editorAttrs && this.options.editorAttrs.disabled) {
-      this.$el.addClass('is-disabled');
-    }
-
-    this._initFillDialog();
-    this._initInputFields();
-
-    this._popupManager = new PopupManager(
-      this._popupConfig.cid,
-      this._popupConfig.$el,
-      this._dialogView.$el
-    );
-  },
-
   _initInputFields: function () {
-    this._inputCollection = new InputCollection();
+    InputFillView.prototype._initInputFields.call(this);
 
     this._initStrokeColorInput();
-
-    this._inputCollection.bind('onInputChanged', this._onInputChanged, this);
   },
 
   _initStrokeColorInput: function () {
@@ -88,53 +43,5 @@ module.exports = CoreView.extend({
 
     this.$('.js-content').append(this._inputStrokeColorView.render().$el);
     this._inputCollection && this._inputCollection.push(this._strokeColorModel);
-  },
-
-  _onInputClick: function (inputModel) {
-    if (inputModel.get('selected')) {
-      this.removeDialog();
-      return;
-    }
-
-    inputModel.set('selected', true);
-    this._dialogView.model.set('createContentView', inputModel.get('createContentView'));
-    this._dialogView.render();
-    this._dialogView.show();
-
-    this._popupManager.append(this._dialogMode);
-    this._popupManager.track();
-  },
-
-  _onInputChanged: function (model) {
-    this.trigger('onInputChanged', this);
-  },
-
-  _initFillDialog: function () {
-    var dialogModel = new DialogModel();
-
-    this.listenToOnce(dialogModel, 'destroy', function () {
-      this._dialogView = null;
-      this.stopListening(dialogModel);
-    });
-
-    this._dialogView = new DialogView({
-      model: dialogModel
-    });
-  },
-
-  removeDialog: function () {
-    this._inputCollection.unselect();
-    this._dialogView.clean();
-    this._popupManager.untrack();
-  },
-
-  focus: function () {
-    if (this.hasFocus) return;
-    this.$('.js-fillInput').focus();
-  },
-
-  blur: function () {
-    if (!this.hasFocus) return;
-    this.$('.js-fillInput').blur();
   }
 });
